freertos_recovid: Add TaskSerialSending
Only one task is allowed to use Serial : TaskSerialSending. The
communication is based on QueueMessage (Frertos object).

The serial task pick the next message and write it to serial

We also provide an helper to write message : SerialPrintf. This function
allows to use a printf formatted way to write to Serial. It will
allocate dynamically