motor.c: before press be sure to be de homed (a little bit)
The main idea is to have the same tension in the rope at each cycle.
We release the rope at the end of insulflation. When home the rope may not have the correct tension
Moving slowly until not home ensure the tension is nearly the same before pressing your new cycle